Use UI-side `SplitTunneling`
